Move mount() to mount.tsx
